[CoreExtension] Reduce duplicate Code
