+++ src/Extension/Core/Type/LocaleType.php
@@ -15,6 +15,7 @@ namespace Rollerworks\Component\Search\Extension\Core\Type;
 
 use Rollerworks\Component\Search\Extension\Core\ChoiceList\ArrayChoiceList;
 use Rollerworks\Component\Search\Extension\Core\ChoiceList\ChoiceList;
+use Rollerworks\Component\Search\Extension\Core\ChoiceList\ChoiceLoaderTrait;
 use Rollerworks\Component\Search\Extension\Core\ChoiceList\Loader\ChoiceLoader;
 use Rollerworks\Component\Search\Field\AbstractFieldType;
 use Symfony\Component\Intl\Intl;
@@ -25,16 +26,7 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
  */
 final class LocaleType extends AbstractFieldType implements ChoiceLoader
 {
-    /**
-     * Locale loaded choice list.
-     *
-     * The choices are lazy loaded and generated from the Intl component.
-     *
-     * {@link \Symfony\Component\Intl\Intl::getLocaleBundle()}.
-     *
-     * @var ArrayChoiceList
-     */
-    private $choiceList;
+    use ChoiceLoaderTrait;
 
     /**
      * {@inheritdoc}
@@ -66,42 +58,6 @@ final class LocaleType extends AbstractFieldType implements ChoiceLoader
         }
 
         return $this->choiceList = new ArrayChoiceList(array_flip(Intl::getLocaleBundle()->getLocaleNames()), $value);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function loadChoicesForValues(array $values, callable $value = null): array
-    {
-        // Optimize
-        if (empty($values)) {
-            return [];
-        }
-
-        // If no callable is set, values are the same as choices
-        if (null === $value) {
-            return $values;
-        }
-
-        return $this->loadChoiceList($value)->getChoicesForValues($values);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function loadValuesForChoices(array $choices, callable $value = null): array
-    {
-        // Optimize
-        if (empty($choices)) {
-            return [];
-        }
-
-        // If no callable is set, choices are the same as values
-        if (null === $value) {
-            return $choices;
-        }
-
-        return $this->loadChoiceList($value)->getValuesForChoices($choices);
     }
 
     /**
